third_party/WebKit/LayoutTests/bluetooth/server/getPrimaryService/service-not-found.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('HeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: ['glucose']}))
    .then(device => device.gatt.connect())
    .then(gattServer => assert_promise_rejects_with_message(
      gattServer.getPrimaryService('glucose'),
      new DOMException('No Services matching UUID ' + glucose.uuid + ' found in Device.',
                       'NotFoundError')));
}, 'Request for absent service. Reject with NotFoundError.');
</script>
